+++ mybooking-templates/mybooking-plugin-complete-tmpl.php
@@ -162,9 +162,13 @@
             <%=shopping_cart.time_to%>
           <% } %> 
         </li>
-        <li class="list-group-item reservation-summary-card-detail"><?php echo _x( 'Rental duration', 'renting_complete', 'mybooking-wp-plugin' ) ?>: <%=shopping_cart.days%> <?php echo _x( 'day(s)', 'renting_complete', 'mybooking-wp-plugin' ) ?></li>
+        <% if (shopping_cart.days > 0) { %>
+          <li class="list-group-item reservation-summary-card-detail"><?php echo _x( 'Rental duration', 'renting_complete', 'mybooking-wp-plugin' ) ?>: <%=shopping_cart.days%> <?php echo _x( 'day(s)', 'renting_complete', 'mybooking-wp-plugin' ) ?></li>
+        <% } else if (shopping_cart.hours > 0) { %>
+          <li class="list-group-item reservation-summary-card-detail"><?php echo _x( 'Rental duration', 'renting_complete', 'mybooking-wp-plugin' ) ?>: <%=shopping_cart.hours%> <?php echo _x( 'hour(s)', 'renting_complete', 'mybooking-wp-plugin' ) ?></li>        
+        <% } %> 
         <li class="list-group-item">
-          <button id="modify_reservation_button" class="btn btn-primary w-100"><?php echo _x( 'Modify reservation', 'renting_complete', 'mybooking-wp-plugin' ) ?></button>
+          <button id="modify_reservation_button" class="btn btn-primary w-100"><?php echo _x( 'Edit', 'renting_complete', 'mybooking-wp-plugin' ) ?></button>
         </li>
       </ul>
     </div>