+++ mybooking-templates/mybooking-plugin-product-widget.php
@@ -1,4 +1,8 @@
-    <div id="product_selector" data-code="<?php echo $args['code']?>" class="container is-desktop">
+    <div id="product_selector" data-code="<?php echo $args['code']?>" 
+                               <?php if ( array_key_exists('sales_channel_code', $args) && $args['sales_channel_code'] != '' ) : ?>
+                               data-sales-channel-code="<?php echo $args['sales_channel_code']?>" 
+                               <?php endif; ?> 
+                               class="container is-desktop">
       <div class="row">
         <div class="col-12">
           <form